Update vacuum.go, vacuum_test.go, TEST1
TEST1 changed in the prompt, so I changed it for Part Two.

In vacuum.go, I changed the house struct, which now includes a string
determining which Santa is delivering the presents. To follow with
this change, processInstructions had to be altered as well, as did
previousHouses so that the houses they had both visited wouldn't
count twice.

I actually used vacuum_test.go this time, and it was a big help.
Holds test functions for the three test files (TEST1, TEST2, TEST3).